Implements Cpanfile parser
